// src/shared/interface/pokemon.interface.d.ts
/**
 * Pokemon Type
 * @author Irfan Andriansyah <irfan@99.co>
 * @since 2020.09.07
 */
export type PokemonType =
    | 'Grass'
    | 'Poison'
    | 'Fire'
    | 'Flying'
    | 'Water'
    | 'Bug'
    | 'Normal'
    | 'Electric'
    | 'Ground'
    | 'Fairy'
    | 'Fighting'
    | 'Psychic'
    | 'Rock'
    | 'Steel'
    | 'Ice'
    | 'Ghost'
    | 'Dragon';